+++ lib/surface/compiler/tokenizer.ex
@@ -14,15 +14,18 @@ defmodule Surface.Compiler.Tokenizer do
     "base",
     "br",
     "col",
+    "command",
+    "embed",
     "hr",
     "img",
     "input",
+    "keygen",
     "link",
     "meta",
     "param",
-    "command",
-    "keygen",
-    "source"
+    "source",
+    "track",
+    "wbr"
   ]
 
   @type metadata :: %{